requirements.txt
Django>=4.2.0,<5.0.0
channels>=4.0.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
ocpp>=0.18.0,<1.0.0
daphne>=4.0.0,<5.0.0
channels_redis>=4.1.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
django-cors-headers>=4.3.0,<5.0.0
djongo>=1.3.6
pymongo>=3.12.3,<4.0.0